CODE RUB: Remove RetrieveByName Logic Tests to Seperated File
